# backend/src/airas/dashboard/api/routes/v1/hypothesis_driven_research.py
import asyncio
import logging
import uuid
from typing import Annotated

from dependency_injector.wiring import Closing, Provide, inject
from fastapi import APIRouter, Depends, HTTPException

from airas.container import Container
from airas.core.types.e2e import Status
from airas.core.types.github import GitHubConfig
from airas.dashboard.api.dependencies import (
    get_current_user_id,
    get_github_client,
    get_github_owner,
    get_litellm_client,
)
from airas.dashboard.api.schemas.hypothesis_driven_research import (
    HypothesisDrivenResearchListItemResponse,
    HypothesisDrivenResearchListResponseBody,
    HypothesisDrivenResearchRequestBody,
    HypothesisDrivenResearchResponseBody,
    HypothesisDrivenResearchStatusResponseBody,
    HypothesisDrivenResearchUpdateRequestBody,
)
from airas.infra.github_client import GithubClient
from airas.infra.litellm_client import LiteLLMClient
from airas.usecases.autonomous_research.e2e_research_service_protocol import (
    E2EResearchServiceProtocol,
)
from airas.usecases.autonomous_research.hypothesis_driven_research.hypothesis_driven_research import (
    HypothesisDrivenResearch,
)

logger = logging.getLogger(__name__)

router = APIRouter(
    prefix="/hypothesis_driven_research", tags=["hypothesis_driven_research"]
)


async def _execute_hypothesis_driven_research(
    task_id: uuid.UUID,
    created_by: uuid.UUID,
    request: HypothesisDrivenResearchRequestBody,
    github_owner: str,
    github_client: GithubClient,
    litellm_client: LiteLLMClient,
    e2e_service: E2EResearchServiceProtocol,
) -> None:
    try:
        logger.info(f"[Task {task_id}] Starting HypothesisDrivenResearch execution")

        graph = HypothesisDrivenResearch(
            github_client=github_client,
            litellm_client=litellm_client,
            e2e_service=e2e_service,
            compute_environment=request.compute_environment,
            runner_config=request.runner_config,
            wandb_config=request.wandb_config,
            task_id=task_id,
            created_by=created_by,
            is_github_repo_private=request.is_github_repo_private,
            num_experiment_models=request.num_experiment_models,
            num_experiment_datasets=request.num_experiment_datasets,
            num_comparison_methods=request.num_comparison_methods,
            paper_content_refinement_iterations=request.paper_content_refinement_iterations,
            github_actions_agent=request.github_actions_agent,
            latex_template_name=request.latex_template_name,
            llm_mapping=request.llm_mapping,
        ).build_graph()

        logger.info(f"[Task {task_id}] Streaming graph execution")

        config = {"recursion_limit": 100}

        async for chunk in graph.astream(
            {
                "task_id": task_id,
                "github_config": GitHubConfig(
                    github_owner=github_owner,
                    repository_name=request.github_config.repository_name,
                    branch_name=request.github_config.branch_name,
                ),
                "research_hypothesis": request.research_hypothesis,
                "research_topic": request.research_topic,
            },
            config=config,
        ):
            for node_name, node_output in chunk.items():
                if not isinstance(node_output, dict):
                    continue

                if "research_history" in node_output:
                    logger.info(
                        f"[Task {task_id}] Research history updated from node: {node_name}"
                    )

    except Exception as e:
        error_msg = f"{type(e).__name__}: {str(e)}"
        logger.exception(f"[Task {task_id}] Execution failed")

        try:
            e2e_service.update(
                id=task_id, status=Status.FAILED, error_message=error_msg
            )
        except Exception:
            logger.exception(
                f"[Task {task_id}] CRITICAL: Failed to update status to FAILED. "
                f"Task may remain in RUNNING state."
            )


@router.post("/run", response_model=HypothesisDrivenResearchResponseBody)
@inject
async def execute_hypothesis_driven_research(
    request: HypothesisDrivenResearchRequestBody,
    current_user_id: Annotated[uuid.UUID, Depends(get_current_user_id)],
    github_owner: Annotated[str, Depends(get_github_owner)],
    github_client: Annotated[GithubClient, Depends(get_github_client)],
    litellm_client: Annotated[LiteLLMClient, Depends(get_litellm_client)],
    e2e_service: Annotated[
        E2EResearchServiceProtocol,
        Depends(Provide[Container.e2e_research_service]),
    ],
) -> HypothesisDrivenResearchResponseBody:
    task_id = uuid.uuid4()

    asyncio.create_task(
        _execute_hypothesis_driven_research(
            task_id=task_id,
            created_by=current_user_id,
            request=request,
            github_owner=github_owner,
            github_client=github_client,
            litellm_client=litellm_client,
            e2e_service=e2e_service,
        )
    )

    return HypothesisDrivenResearchResponseBody(task_id=task_id)


@router.get(
    "/status/{task_id}", response_model=HypothesisDrivenResearchStatusResponseBody
)
@inject
async def get_hypothesis_driven_research_status(
    task_id: uuid.UUID,
    e2e_service: Annotated[
        E2EResearchServiceProtocol,
        Depends(Closing[Provide[Container.e2e_research_service]]),
    ],
) -> HypothesisDrivenResearchStatusResponseBody:
    try:
        result = e2e_service.get(task_id)
    except ValueError as exc:
        raise HTTPException(
            status_code=404, detail=f"Task {task_id} not found"
        ) from exc
    except Exception as exc:  # pragma: no cover - defensive
        logger.exception(f"[Status {task_id}] Failed to get status")
        raise HTTPException(status_code=500, detail="Internal Server Error") from exc

    try:
        return HypothesisDrivenResearchStatusResponseBody.model_validate(result)
    except Exception as exc:  # pragma: no cover - defensive
        logger.exception(f"[Status {task_id}] Failed to validate response")
        raise HTTPException(
            status_code=500, detail="Failed to validate response"
        ) from exc


@router.get("", response_model=HypothesisDrivenResearchListResponseBody)
@inject
async def list_hypothesis_driven_research(
    e2e_service: Annotated[
        E2EResearchServiceProtocol,
        Depends(Closing[Provide[Container.e2e_research_service]]),
    ],
    offset: int = 0,
    limit: int | None = None,
) -> HypothesisDrivenResearchListResponseBody:
    records = e2e_service.list(offset=offset, limit=limit)
    sorted_records = sorted(records, key=lambda record: record.created_at, reverse=True)

    return HypothesisDrivenResearchListResponseBody(
        items=[
            HypothesisDrivenResearchListItemResponse.model_validate(record)
            for record in sorted_records
        ]
    )


@router.patch(
    "/{task_id}",
    response_model=HypothesisDrivenResearchStatusResponseBody,
    status_code=200,
)
@inject
async def update_hypothesis_driven_research(
    task_id: uuid.UUID,
    request: HypothesisDrivenResearchUpdateRequestBody,
    e2e_service: Annotated[
        E2EResearchServiceProtocol,
        Depends(Closing[Provide[Container.e2e_research_service]]),
    ],
) -> HypothesisDrivenResearchStatusResponseBody:
    updates = request.model_dump(exclude_none=True)
    if not updates:
        raise HTTPException(status_code=400, detail="No fields provided for update")

    try:
        updated = e2e_service.update(id=task_id, **updates)
    except ValueError as exc:
        raise HTTPException(status_code=404, detail=str(exc)) from exc
    except Exception as exc:  # pragma: no cover - defensive
        logger.exception(f"[Update {task_id}] Failed to update")
        raise HTTPException(status_code=500, detail="Internal Server Error") from exc

    try:
        return HypothesisDrivenResearchStatusResponseBody.model_validate(updated)
    except Exception as exc:  # pragma: no cover - defensive
        logger.exception(f"[Update {task_id}] Failed to validate response")
        raise HTTPException(
            status_code=500, detail="Failed to validate response"
        ) from exc