# schema/openapi.yaml
openapi: 3.1.0
info:
  title: AIRAS API
  version: 0.0.1
paths:
...
  /airas/v1/papers/search:
    post:
      tags:
      - papers
      summary: Search Paper Titles
      operationId: search_paper_titles_airas_v1_papers_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchPaperTitlesRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchPaperTitlesResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /airas/v1/interactive-repo-agent/dispatch:
    post:
      tags:
      - interactive-repo-agent
      summary: Dispatch Interactive Repo Agent
      operationId: dispatch_interactive_repo_agent_airas_v1_interactive_repo_agent_dispatch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchInteractiveRepoAgentRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchInteractiveRepoAgentResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/interactive-repo-agent/{workflow_run_id}/cancel:
    post:
      tags:
      - interactive-repo-agent
      summary: Cancel Interactive Repo Agent
      operationId: cancel_interactive_repo_agent_airas_v1_interactive_repo_agent__workflow_run_id__cancel_post
      parameters:
      - name: workflow_run_id
        in: path
        required: true
        schema:
          type: integer
          title: Workflow Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelInteractiveRepoAgentRequestBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelInteractiveRepoAgentResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /airas/v1/hypothesis_driven_research/run:
    post:
      tags:
      - hypothesis_driven_research
      summary: Execute Hypothesis Driven Research
      operationId: execute_hypothesis_driven_research_airas_v1_hypothesis_driven_research_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HypothesisDrivenResearchRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypothesisDrivenResearchResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/hypothesis_driven_research/status/{task_id}:
    get:
      tags:
      - hypothesis_driven_research
      summary: Get Hypothesis Driven Research Status
      operationId: get_hypothesis_driven_research_status_airas_v1_hypothesis_driven_research_status__task_id__get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypothesisDrivenResearchStatusResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/hypothesis_driven_research:
    get:
      tags:
      - hypothesis_driven_research
      summary: List Hypothesis Driven Research
      operationId: list_hypothesis_driven_research_airas_v1_hypothesis_driven_research_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypothesisDrivenResearchListResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /airas/v1/hypothesis_driven_research/{task_id}:
    patch:
      tags:
      - hypothesis_driven_research
      summary: Update Hypothesis Driven Research
      operationId: update_hypothesis_driven_research_airas_v1_hypothesis_driven_research__task_id__patch
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HypothesisDrivenResearchUpdateRequestBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HypothesisDrivenResearchStatusResponseBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CancelInteractiveRepoAgentRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
      type: object
      required:
      - github_config
      title: CancelInteractiveRepoAgentRequestBody
    CancelInteractiveRepoAgentResponseBody:
      properties:
        cancelled:
          type: boolean
          title: Cancelled
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - cancelled
      - execution_time
      title: CancelInteractiveRepoAgentResponseBody
...
    ComputeEnvironment:
      properties:
        os:
          anyOf:
          - type: string
          - type: 'null'
          title: Os
        cpu_cores:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cpu Cores
        ram_gb:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ram Gb
        gpu_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Gpu Type
        gpu_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Gpu Count
        gpu_memory_gb:
          anyOf:
          - type: integer
          - type: 'null'
          title: Gpu Memory Gb
        storage_type:
          anyOf:
          - type: string
            enum:
            - nvme
            - ssd
            - hdd
          - type: 'null'
          title: Storage Type
        storage_gb:
          anyOf:
          - type: integer
          - type: 'null'
          title: Storage Gb
        python_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Python Version
        cuda_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Cuda Version
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      title: ComputeEnvironment
...
    DispatchInteractiveRepoAgentRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfig'
        github_actions_agent:
          type: string
          enum:
          - claude_code
          - open_code
          title: Github Actions Agent
        session_username:
          type: string
          title: Session Username
        session_password:
          type: string
          format: password
          title: Session Password
          writeOnly: true
      type: object
      required:
      - github_config
      - github_actions_agent
      - session_username
      - session_password
      title: DispatchInteractiveRepoAgentRequestBody
    DispatchInteractiveRepoAgentResponseBody:
      properties:
        dispatched:
          type: boolean
          title: Dispatched
        workflow_run_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Workflow Run Id
        tunnel_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Tunnel Url
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - dispatched
      - workflow_run_id
      - tunnel_url
      - execution_time
      title: DispatchInteractiveRepoAgentResponseBody
...
    EphemeralCloudRunnerConfig:
      properties:
        type:
          type: string
          const: ephemeral_cloud
          title: Type
        cloud_provider:
          type: string
          enum:
          - aws
          - gcp
          title: Cloud Provider
          description: Cloud provider ('aws' or 'gcp')
          default: aws
        gpu_instance_type:
          type: string
          title: Gpu Instance Type
          description: Instance type (e.g., 'g4dn.xlarge' for AWS, 'n1-standard-4'
            for GCP)
          default: g4dn.xlarge
        max_instance_hours:
          type: integer
          minimum: 1.0
          title: Max Instance Hours
          description: Max instance lifetime in hours (safety net for orphaned instances)
          default: 120
      type: object
      required:
      - type
      title: EphemeralCloudRunnerConfig
...
    ExperimentCycleGraphLLMMapping:
      properties:
        code_generation:
          anyOf:
          - $ref: '#/components/schemas/CodeGenerationGraphLLMMapping'
          - type: 'null'
        dispatch_experiment_validation:
          anyOf:
          - $ref: '#/components/schemas/DispatchExperimentValidationLLMMapping'
          - type: 'null'
        analyze_experiment:
          anyOf:
          - $ref: '#/components/schemas/AnalyzeExperimentLLMMapping'
          - type: 'null'
        decide_experiment_cycle:
          anyOf:
          - $ref: '#/components/schemas/DecideExperimentCycleLLMMapping'
          - type: 'null'
        refine_experimental_design:
          anyOf:
          - $ref: '#/components/schemas/RefineExperimentalDesignLLMMapping'
          - type: 'null'
      type: object
      title: ExperimentCycleGraphLLMMapping
...
    GenerateExperimentalDesignLLMMapping:
      properties:
        generate_experimental_design:
          $ref: '#/components/schemas/NodeLLMConfig'
      type: object
      required:
      - generate_experimental_design
      title: GenerateExperimentalDesignLLMMapping
...
    GitHubConfig:
      properties:
        github_owner:
          type: string
          title: Github Owner
          description: GitHub owner of the repository
        repository_name:
          type: string
          title: Repository Name
          description: Name of the repository
        branch_name:
          type: string
          title: Branch Name
          description: Branch name
      type: object
      required:
      - github_owner
      - repository_name
      - branch_name
      title: GitHubConfig
    GitHubConfigRequest:
      properties:
        repository_name:
          type: string
          title: Repository Name
          description: Name of the repository
        branch_name:
          type: string
          title: Branch Name
          description: Branch name
      type: object
      required:
      - repository_name
      - branch_name
      title: GitHubConfigRequest
      description: Request body schema for GitHub config (owner is resolved server-side).
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HypothesisDrivenResearchLLMMapping:
      properties:
        generate_experimental_design:
          anyOf:
          - $ref: '#/components/schemas/GenerateExperimentalDesignLLMMapping'
          - type: 'null'
        experiment_cycle:
          anyOf:
          - $ref: '#/components/schemas/ExperimentCycleGraphLLMMapping'
          - type: 'null'
        write:
          anyOf:
          - $ref: '#/components/schemas/WriteLLMMapping'
          - type: 'null'
        latex:
          anyOf:
          - $ref: '#/components/schemas/LaTeXGraphLLMMapping'
          - type: 'null'
      type: object
      title: HypothesisDrivenResearchLLMMapping
    HypothesisDrivenResearchListItemResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        created_by:
          type: string
          format: uuid
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        status:
          $ref: '#/components/schemas/Status'
        current_step:
          anyOf:
          - $ref: '#/components/schemas/StepType'
          - type: 'null'
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        last_updated_at:
          type: string
          format: date-time
          title: Last Updated At
        github_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Github Url
      type: object
      required:
      - id
      - title
      - created_by
      - created_at
      - status
      - last_updated_at
      title: HypothesisDrivenResearchListItemResponse
    HypothesisDrivenResearchListResponseBody:
      properties:
        items:
          items:
            $ref: '#/components/schemas/HypothesisDrivenResearchListItemResponse'
          type: array
          title: Items
      type: object
      required:
      - items
      title: HypothesisDrivenResearchListResponseBody
    HypothesisDrivenResearchRequestBody:
      properties:
        github_config:
          $ref: '#/components/schemas/GitHubConfigRequest'
        research_hypothesis:
          $ref: '#/components/schemas/ResearchHypothesis'
        research_topic:
          type: string
          title: Research Topic
          default: ''
        compute_environment:
          $ref: '#/components/schemas/ComputeEnvironment'
        runner_config:
          oneOf:
          - $ref: '#/components/schemas/StaticRunnerConfig'
          - $ref: '#/components/schemas/EphemeralCloudRunnerConfig'
          title: Runner Config
          discriminator:
            propertyName: type
            mapping:
              ephemeral_cloud: '#/components/schemas/EphemeralCloudRunnerConfig'
              static: '#/components/schemas/StaticRunnerConfig'
        wandb_config:
          $ref: '#/components/schemas/WandbConfig'
        is_github_repo_private:
          type: boolean
          title: Is Github Repo Private
          default: false
        num_experiment_models:
          type: integer
          title: Num Experiment Models
          default: 1
        num_experiment_datasets:
          type: integer
          title: Num Experiment Datasets
          default: 1
        num_comparison_methods:
          type: integer
          title: Num Comparison Methods
          default: 1
        paper_content_refinement_iterations:
          type: integer
          title: Paper Content Refinement Iterations
          default: 2
        github_actions_agent:
          type: string
          enum:
          - claude_code
          - open_code
          title: Github Actions Agent
          default: open_code
        latex_template_name:
          type: string
          title: Latex Template Name
          default: mdpi
        llm_mapping:
          anyOf:
          - $ref: '#/components/schemas/HypothesisDrivenResearchLLMMapping'
          - type: 'null'
      type: object
      required:
      - github_config
      - research_hypothesis
      - compute_environment
      - runner_config
      - wandb_config
      title: HypothesisDrivenResearchRequestBody
    HypothesisDrivenResearchResponseBody:
      properties:
        task_id:
          type: string
          format: uuid
          title: Task Id
      type: object
      required:
      - task_id
      title: HypothesisDrivenResearchResponseBody
    HypothesisDrivenResearchStatusResponseBody:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        created_by:
          type: string
          format: uuid
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        status:
          $ref: '#/components/schemas/Status'
        current_step:
          anyOf:
          - $ref: '#/components/schemas/StepType'
          - type: 'null'
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        last_updated_at:
          type: string
          format: date-time
          title: Last Updated At
        result:
          additionalProperties: true
          type: object
          title: Result
        github_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Github Url
        task_id:
          type: string
          format: uuid
          title: Task Id
          readOnly: true
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
          readOnly: true
        research_history:
          anyOf:
          - $ref: '#/components/schemas/ResearchHistory-Output'
          - type: 'null'
          readOnly: true
      type: object
      required:
      - id
      - title
      - created_by
      - created_at
      - status
      - last_updated_at
      - result
      - task_id
      - error
      - research_history
      title: HypothesisDrivenResearchStatusResponseBody
    HypothesisDrivenResearchUpdateRequestBody:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        status:
          anyOf:
          - $ref: '#/components/schemas/Status'
          - type: 'null'
        current_step:
          anyOf:
          - $ref: '#/components/schemas/StepType'
          - type: 'null'
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
        github_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Github Url
      type: object
      title: HypothesisDrivenResearchUpdateRequestBody
...
    ResearchHypothesis:
      properties:
        open_problems:
          type: string
          title: Open Problems
        method:
          type: string
          title: Method
        experimental_setup:
          type: string
          title: Experimental Setup
        primary_metric:
          type: string
          title: Primary Metric
        experimental_code:
          type: string
          title: Experimental Code
        expected_result:
          type: string
          title: Expected Result
        expected_conclusion:
          type: string
          title: Expected Conclusion
      type: object
      required:
      - open_problems
      - method
      - experimental_setup
      - primary_metric
      - experimental_code
      - expected_result
      - expected_conclusion
      title: ResearchHypothesis
...
    SearchPaperTitlesRequestBody:
      properties:
        search_method:
          type: string
          enum:
          - airas_db
          - qdrant
          title: Search Method
          default: airas_db
        queries:
          items:
            type: string
          type: array
          title: Queries
        max_results_per_query:
          type: integer
          exclusiveMinimum: 0.0
          title: Max Results Per Query
          default: 3
        collection_name:
          type: string
          title: Collection Name
          default: airas_papers_db
      type: object
      required:
      - queries
      title: SearchPaperTitlesRequestBody
    SearchPaperTitlesResponseBody:
      properties:
        paper_titles:
          items:
            type: string
          type: array
          title: Paper Titles
        execution_time:
          additionalProperties:
            items:
              type: number
            type: array
          type: object
          title: Execution Time
      type: object
      required:
      - paper_titles
      - execution_time
      title: SearchPaperTitlesResponseBody
...
    StaticRunnerConfig:
      properties:
        type:
          type: string
          const: static
          title: Type
        runner_label:
          items:
            type: string
          type: array
          title: Runner Label
          description: Runner labels used by GitHub Actions (e.g., ['ubuntu-latest']
            or ['self-hosted', 'gpu-runner'])
      type: object
      required:
      - type
      title: StaticRunnerConfig
    Status:
      type: string
      enum:
      - pending
      - running
      - completed
      - failed
      title: Status
    StepType:
      type: string
      enum:
      - generate_queries
      - search_paper_titles
      - retrieve_papers
      - generate_hypothesis
      - generate_experimental_design
      - generate_code
      - push_code
      - execute_trial_experiment
      - execute_full_experiment
      - execute_evaluation_workflow
      - fetch_experiment_results
      - analyze_experiment_results
      - generate_bibfile
      - generate_paper
      - generate_latex
      - compile_latex
      - done
      title: StepType
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WandbConfig:
      properties:
        entity:
          type: string
          title: Entity
          description: Wandb entity (username or team name)
        project:
          type: string
          title: Project
          description: Wandb project name
        run_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Run Ids
          description: List of Wandb run IDs (optional, can be retrieved from metadata)
      type: object
      required:
      - entity
      - project
      title: WandbConfig
    WriteLLMMapping:
      properties:
        write_paper:
          $ref: '#/components/schemas/NodeLLMConfig'
        refine_paper:
          $ref: '#/components/schemas/NodeLLMConfig'
      type: object
      required:
      - write_paper
      - refine_paper
      title: WriteLLMMapping
...